배경택:[Add] 상점 UI 구현 완료
