--- Metalord_btin/MetaLord/Assets/_Test/BKT/Scripts/Record/Utility.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ec8a1b3bac64747c9a7f770d8ed8c2d8
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
