--- Metalord_btin/MetaLord/Assets/_Test/BKT/Scripts/Store/UI_StoreCoin.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ac11d9eda807b4c5d93d66cc970bf34e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
